+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [UNRELEASED]
+
+### Added
+- Support for set labels on AssetUpdate
+
 ## [2.25.0] - 2021-07-06
 
 ### Added